pyproject.toml
[project]
name = "journal"
version = "0.1.0"
description = "Personal journal insight engine — ingest handwritten pages and voice notes, query with natural language"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86,<1",
    "openai>=2.29,<3",
    "chromadb-client>=1.5,<2",
    "mcp[cli]>=1.26,<2",
    "tiktoken>=0.9,<1",
]

[project.scripts]
journal = "journal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/journal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/journal"]

[tool.coverage.report]
show_missing = true
fail_under = 65

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.7",
]
